Adds the concept for a Heatmap Prop generator
