+++ generator/src/main/java/com/faforever/neroxis/generator/prop/BasicPropGenerator.java
@@ -12,6 +12,7 @@ public class BasicPropGenerator extends PropGenerator {
     protected BooleanMask treeMask;
     protected BooleanMask cliffRockMask;
     protected BooleanMask fieldStoneMask;
+    protected BooleanMask noProps;
 
     @Override
     public void initialize(SCMap map, long seed, GeneratorParameters generatorParameters,
@@ -20,23 +21,22 @@ public class BasicPropGenerator extends PropGenerator {
         treeMask = new BooleanMask(1, random.nextLong(), symmetrySettings, "treeMask");
         cliffRockMask = new BooleanMask(1, random.nextLong(), symmetrySettings, "cliffRockMask");
         fieldStoneMask = new BooleanMask(1, random.nextLong(), symmetrySettings, "fieldStoneMask");
+        noProps = new BooleanMask(1, null, symmetrySettings, "noProps");
     }
 
     @Override
     public void placeProps() {
-        placePropsWithExclusion(generatePropExclusionMasks());
+        generatePropExclusionMasks();
+        placePropsWithExclusion();
     }
 
     @Override
     public void placeUnits() {
     }
 
-    protected BooleanMask generatePropExclusionMasks() {
-        BooleanMask noProps = new BooleanMask(1, null, symmetrySettings, "noProps");
+    protected void generatePropExclusionMasks() {
         noProps.init(unbuildable.getFinalMask());
-
         generateExclusionZones(noProps, 30, 2, 8);
-        return noProps;
     }
 
     protected void generateExclusionZones(BooleanMask mask, float spawnSpacing, float mexSpacing, float hydroSpacing) {
@@ -45,7 +45,7 @@ public class BasicPropGenerator extends PropGenerator {
         map.getHydros().forEach(hydro -> mask.fillCircle(hydro.getPosition(), hydroSpacing, true));
     }
 
-    public void placePropsWithExclusion(BooleanMask noProps) {
+    public void placePropsWithExclusion() {
         DebugUtil.timedRun("com.faforever.neroxis.map.generator", "placeProps", () -> {
             Biome biome = map.getBiome();
             propPlacer.placeProps(treeMask.getFinalMask().subtract(noProps), biome.propMaterials().treeGroups(),
