+++ generator/src/main/java/com/faforever/neroxis/generator/prop/BoulderFieldPropGenerator.java
@@ -25,7 +25,7 @@ public class BoulderFieldPropGenerator extends BasicPropGenerator {
     }
 
     @Override
-    public void placePropsWithExclusion(BooleanMask noProps) {
+    public void placePropsWithExclusion() {
         DebugUtil.timedRun("com.faforever.neroxis.map.generator", "placeProps", () -> {
             Biome biome = map.getBiome();
             propPlacer.placeProps(treeMask.getFinalMask().subtract(noProps), biome.propMaterials().treeGroups(),
