+++ generator/src/main/java/com/faforever/neroxis/generator/prop/NeutralCivPropGenerator.java
@@ -44,8 +44,9 @@ public class NeutralCivPropGenerator extends BasicPropGenerator {
     }
 
     @Override
-    protected BooleanMask generatePropExclusionMasks() {
-        return super.generatePropExclusionMasks().add(civReclaimMask.getFinalMask());
+    protected void generatePropExclusionMasks() {
+        super.generatePropExclusionMasks();
+        noProps.add(civReclaimMask.getFinalMask());
     }
 
     @Override
