+++ generator/src/main/java/com/faforever/neroxis/generator/prop/ReducedNaturalPropGenerator.java
@@ -6,7 +6,7 @@ import com.faforever.neroxis.util.DebugUtil;
 
 public abstract class ReducedNaturalPropGenerator extends BasicPropGenerator {
     @Override
-    public void placePropsWithExclusion(BooleanMask noProps) {
+    public void placePropsWithExclusion() {
         DebugUtil.timedRun("com.faforever.neroxis.map.generator", "placeProps", () -> {
             Biome biome = map.getBiome();
             propPlacer.placeProps(treeMask.getFinalMask().subtract(noProps), biome.propMaterials().treeGroups(),
