+++ generator/src/main/java/com/faforever/neroxis/generator/prop/RockFieldPropGenerator.java
@@ -19,7 +19,7 @@ public class RockFieldPropGenerator extends BasicPropGenerator {
     }
 
     @Override
-    public void placePropsWithExclusion(BooleanMask noProps) {
+    public void placePropsWithExclusion() {
         DebugUtil.timedRun("com.faforever.neroxis.map.generator", "placeProps", () -> {
             Biome biome = map.getBiome();
             propPlacer.placeProps(treeMask.getFinalMask().subtract(noProps), biome.propMaterials().treeGroups(),
